requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.12.1
idna==3.4
llvmlite==0.39.1
more-itertools==9.1.0
multidict==6.0.4
numba==0.56.4
numpy==1.23.5
openai==0.27.0
openai-whisper @ git+https://github.com/openai/whisper.git@aac47c98349b98cec5ca7b1be53960fb59f4436b
packaging==23.0
PyAudio==0.2.13
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
simpleaudio==1.0.4
six==1.16.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2